add menu and map to storeDetail.jsp

+++ src/main/java/com/sw/fd/repository/StoreRepository.java
@@ -4,8 +4,9 @@ import com.sw.fd.entity.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
 import java.util.Optional;
 
-public interface StoreRepository extends JpaRepository<Store, Integer> {
+public interface StoreRepository extends JpaRepository<Store, Long> {
     Optional<Store> findBySno(int sno);
 }
